backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
vaderSentiment>=3.3.2
numpy>=1.24.0
gunicorn>=21.0.0
openai>=1.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
